requirements.txt
torch
transformers
accelerate
bitsandbytes
langchain
sentence-transformers
faiss-cpu
openpyxl
pacmap
python-decouple
pandas
datasets
matplotlib
langchain
ragatouille
